// src/data/generated/galleryManifest.ts
export interface GeneratedGalleryManifest {
  branches: Record<string, string[]>;
  storeFrontThumbnails: Record<string, string>;
  seminars: string[];
  events: string[];
  certificates: string[];
}

export const galleryManifest: GeneratedGalleryManifest = {
  "branches": {
    "bhiwandi-kongaon-1": [
      "branch-photos/bhiwandi-kongaon-1/Kongaon Store Front.jpg"
    ],
    "diva": [
      "branch-photos/Diva/Diva 1.jpg"
    ],
    "dombivli-ramnagar": [
      "branch-photos/dombivli-ramnagar/enterance img .jpg",
      "branch-photos/dombivli-ramnagar/Ramnagar 1.jpg",
      "branch-photos/dombivli-ramnagar/Ramnagar 2.jpg"
    ],
    "dombivli-star-colony": [
      "branch-photos/dombivli-star-colony/Star Colony 1.jpg",
      "branch-photos/dombivli-star-colony/Star Colony 2.jpg",
      "branch-photos/dombivli-star-colony/Star Colony 3.jpg",
      "branch-photos/dombivli-star-colony/Star Colony 4.jpg",
      "branch-photos/dombivli-star-colony/Star Colony 5.jpg",
      "branch-photos/dombivli-star-colony/Star Colony 6.jpg"
    ],
    "kalyan-chakkinaka": [
      "branch-photos/kalyan-chakkinaka/Front.jpg",
      "branch-photos/kalyan-chakkinaka/Inside .jpg"
    ],
    "kalyan-rambaug": [
      "branch-photos/kalyan-rambaug/Rambaug-Shop-Photo.jpeg"
    ],
    "kalyan-tilak-chowk": [
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 1.jpg",
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 2.jpg",
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 3.jpg",
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 4.jpg",
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 5.jpg",
      "branch-photos/kalyan-tilak-chowk/Tilak Chowk 6.jpg"
    ],
    "thane-vartak-nagar": [
      "branch-photos/thane-vartak-nagar/1.jpg",
      "branch-photos/thane-vartak-nagar/2.jpg",
      "branch-photos/thane-vartak-nagar/3.jpg",
      "branch-photos/thane-vartak-nagar/4.jpg",
      "branch-photos/thane-vartak-nagar/5.jpg",
      "branch-photos/thane-vartak-nagar/Store Front.jpg"
    ]
  },
  "storeFrontThumbnails": {
    "bhiwandi-kongaon-1": "store-front-thumbnails/bhiwandi-kongaon-1/Kongaon Store Front.jpg",
    "diva": "store-front-thumbnails/diva/Diva 1.jpg",
    "dombivli-ramnagar": "store-front-thumbnails/dombivli-ramnagar/enterance img .jpg",
    "dombivli-star-colony": "store-front-thumbnails/dombivli-star-colony/Star Colony.jpg",
    "kalyan-chakkinaka": "store-front-thumbnails/kalyan-chakkinaka/Front.jpg",
    "kalyan-khadakpada": "store-front-thumbnails/kalyan-khadakpada/IMG-20250730-WA0009.jpg",
    "kalyan-rambaug": "store-front-thumbnails/kalyan-rambaug/Rambaug-Shop-Photo.jpeg",
    "kalyan-tilak-chowk": "store-front-thumbnails/kalyan-tilak-chowk/1.jpg",
    "thane-vartak-nagar": "store-front-thumbnails/thane-vartak-nagar/Store Front.jpg"
  },
  "seminars": [
    "gallery/seminars/Kongaon 1.jpeg",
    "gallery/seminars/Kongaon 2.jpeg",
    "gallery/seminars/Kongaon 3.jpeg",
    "gallery/seminars/Kongaon 4.jpeg",
    "gallery/seminars/Kongaon 5.jpeg",
    "gallery/seminars/Kongaon 6.jpeg",
    "gallery/seminars/Tilak Chowk 1.jpg",
    "gallery/seminars/Tilak Chowk 2.jpg",
    "gallery/seminars/Tilak Chowk 3.jpg",
    "gallery/seminars/Tilak Chowk 4.jpg",
    "gallery/seminars/Tilak Chowk 5.jpg",
    "gallery/seminars/Tilak Chowk 6.jpg",
    "gallery/seminars/Tilak Chowk 7.jpg",
    "gallery/seminars/Tilak Chowk 8.jpg",
    "gallery/seminars/Tilak Chowk 9.jpg",
    "gallery/seminars/Tilak Chowk 10.jpg",
    "gallery/seminars/Tilak Chowk 11.jpg"
  ],
  "events": [
    "gallery/events/ganpati/Star Colony 1.jpg",
    "gallery/events/picnic/Tilak Chowk 1.jpg",
    "gallery/events/picnic/Tilak Chowk 2.jpg",
    "gallery/events/picnic/Tilak Chowk 3.jpg",
    "gallery/events/picnic/Tilak Chowk 4.jpg",
    "gallery/events/spark/Kongaon 4.JPG",
    "gallery/events/spark/Kongaon 5.JPG",
    "gallery/events/spark/Kongaon 6.JPG",
    "gallery/events/spark/Kongaon 7.JPG",
    "gallery/events/spark/Star Colony 3.jpg",
    "gallery/events/staff-sports/bhiwandi-kongaon-1/Kongaon 1.jpeg",
    "gallery/events/staff-sports/bhiwandi-kongaon-1/Kongaon 2.jpeg",
    "gallery/events/staff-sports/bhiwandi-kongaon-1/Kongaon 3.jpeg",
    "gallery/events/staff-sports/dombivli-star-colony/Star Colony 2.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 5.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 6.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 7.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 8.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 10.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 11.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 12.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 13.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 14.jpg",
    "gallery/events/staff-sports/kalyan-tilak-chowk/Tilak Chowk 16.jpg",
    "gallery/events/staff-sports/Tilak Chowk 9.jpg",
    "gallery/events/staff-sports/Tilak Chowk 15.jpg"
  ],
  "certificates": []
};